+++ src/components/NewPostModal/index.tsx
@@ -1,0 +1,7 @@
+import Presenter from './presenter';
+
+const NewPostModal = () => {
+  return <Presenter />;
+};
+
+export default NewPostModal;
